Fix bug with repeated new workspace or error notifications
When the view switches away from the extension and then back again,
the providers remain in memory, including their loaded states and
clusters, but the View component gets unloaded. When the extension
is brought back into view (note the extension is never activated
or deactivated, just removed from view and then brought back in),
the View component gets rendered again, but with reset state,
and because the providers remained in memory, the View thinks
clusters were just added to Lens and shows the notifications (or
thinks an error occurred and shows the error).

This moves the responsibility of posting errors and showing new
workspace notifications into the respective providers, which fixes
the issue.

Also, new workspace notifications now use rich text and show the
extension's name so that users better know where they come from.