Remove Husky and post-merge hook
Husky v5 has gone weird. It now needs to be auto-installed, and
because of that, you have to use `pinst` to temporarily disable
the `postinstall` script in package.json when the package gets
published so it gets published with a `_postinstall` in package.json
that effectively disables the auto-husky install for consumers.

Clunky. And not adding much value. And they changed the license
from MIT to some "Third Party" proprietary license. They say
it'll change back to MIT "later", but feels strange.