[#249] Work toward using the new Lens 5 Catalog to add clusters
TODO:
- Address `TODO[catalog]` comments in the code since clusters
  are being created, but they aren't showing-up in the Catalog
- BLOCKED: https://github.com/lensapp/lens/pull/2775
- BLOCKED: https://github.com/lensapp/lens/pull/2845

R.I.P. workspaces!

MAJOR RE-ORG OF FILES to split them up into `main` and `renderer`
branches so it's clearer what code is meant to run on what
process, and where the IPC divide lies in the code base.

At this point `/src/main` is akin to the __backend__ and
`/src/renderer` is akin to the __frontend__. The IPC API must be
used to communicate between the two since they represent different
processes in Electron.