[#359] WIP: Allow multi-cluster selection with tokenless kubeConfigs
NOTE: This branch is based on the v300 branch.

Requires Lens to install the https://github.com/int128/kubelogin
binary so that fetching the tokens is postponed to the moment
when the cluster is activated in Lens, rather than the time when
the kubeConfig is generated, which means we don't have to go to
the browser for every cluster to get its unique tokens.

TODO:
- See `// DEBUG` comments in the code
- Create a separate template function for the 'exec' style vs
  'auth-provider' style that needs tokens
- Preserve existing SSO-to-browser flow, just allow multiple
  clusters to be selected and immediately call _ssoFinishAddClusters()
  without waiting for the browser round-trip
- Add to docs info to expect browser to be activated when they open
  a cluster for which their kubeclt client doesn't yet have tokens