Adding SshKey, Credential, Proxy Catalog entities (#628)
Adds custom Catalog categories and entities for SSH Keys, Credentials, and Proxies.

Fake items are generated in the Catalog in Dev builds only, just to show how it all works
and provide a template for future work when we'll need these entities (and possibly
others).

Each entity has a custom catalog menu, context menu, and details panel.

Also the native cluster category gets a custom catalog menu item for
creating new clusters.

None the actions do anything other than log to the console for now. We'll
fill this in once we get to implementing those features.