Add status change event dispatching to Cloud
I also fixed a few other issues I found, including how some of
the new Cloud properties were not setting the `_changed` flag.
I did that for consistency with the rest of the code, but I don't
think we need that property anymore, so we could clean that up.

It's only used in one place elsewhere in the code, and that place
could just assume it's always changed.