[PRODX-21176] Enable opening clusters in Lens based on context (#692)
When manually adding a cluster from MCC into Lens via downloaded
kubeconfig, Lens generates its own UID for the cluster since the
UID isn't given in the kubeconfig. That means we can't search
by UID when opening a cluster from MCC in the browser that
wasn't added via the extension.

This changes to searching by the same context name that MCC generates
in its kubeconfigs. So now we can find any cluster, whether it was
manually added, or added via the extension.