Fix CloudStore write to disk issue
This stemed from the optimization I tried to add in
`CloudStore.fromStore()` where I was ignoring new `store.clouds`
values once `this.clouds` was set.

Problem is, CloudStore actually exists on two threads: Main and Renderer.
Lens keeps them in sync via IPC in case one thread changes `this.clouds`.

So we have to go to greater lengths in `fromStore()` to add/update/remove
Clouds in `this.clouds`.