Use existing Namespace class in ExtededCloud instead of anonymous object (#717)
We have a Namespace, and a Cluster class, yet `ExtendedCloud.namespaces`
was a list of anonymous objects with properties.

This moves it to being a list of Namespace objects using the existing
Namespace class, adding necessary properties to the class for clusters,
credentials, and SSH keys.

We're currently missing classes for Credentials and SSH keys. We'll need
to add those in a future PR. I will create a ticket for that(*). We should
propagate the use of raw API objects throughout the code.

This also changes the ExtendedCloud to always fetch ALL namespaces instead
of just the synced ones since we need to know what all exists at all times,
and adds a convenience `syncedNamespaces` getter that filters that list
to just what the Cloud is configured to sync for purposes of UI display
and (future) SyncManager.

* https://mirantis.jira.com/browse/PRODX-21692