Consolidate Cloud.isConnected() and Cloud.status, fix connected notion
I noticed that we ended-up with 2 ways of determining if a Cloud is
"connected" (which will lead to bugs down the line), and also an
incorrect notion of what "connected" means.

Now, we have a single way, `Cloud.status`, and a helper with
`Cloud.connected` (a boolean getter instead of a function, which
is true if status === CONNECTED, false otherwise, for cases where
the binary state is all we care about). Both properties are now
derived from the same set of logic in the `Cloud.status` getter,
so they won't diverge.

Also, "connected" really means the Cloud has all the necessary info
in order to make an API request. Without the `config` loaded, we're
technically not connected yet. So now, if we have everything except
the config, we consider this "connecting" because we have tokens,
but we just haven't gotten the config yet.

This covers the case where the Cloud is restored from disk with valid
tokens as the Cloud doesn't auto-fetch its config. The ExtendedCloud
will fetch the config when it does its initial data fetch. Then the
Cloud will become fully "connected".

I also tested some error scenarios in `Cloud.connect()` and ended-up
tweaking how `ExtendedCloud.fetchData()` reacts as a result. I think
things are more robust now than they were before.