Remove NS "sync now", fix "onClick not a function" bug
"Sync now" doesn't belong on the namespace level.

Also, there was an error in the console when toggling a context menu
open/closed: "onClick is not a function". This was due to passing
a boolean to the `<MenuActions>` `onOpen` prop instead of a function.

But it wasn't necessary to do anything on open, and also the related
state variable wasn't necessary either.