Reconcile custom Catalog entities with API Objects, add License
This update adds the missing LicenseEntity for adding licenses to the Catalog.

It also establishes all the typesets for expected shapes of catalog entities
of all the custom types.

It removes all the creation-related context menu items since those aren't
part of MVP (except maybe "create cluster", but let's see if we get there).

And finally, it adds a `toEntity()` method to each of the API types
(Cluster, Credential, Proxy, SshKey, License, Namespace) to use when
adding items to the Catalog:

```
catalogSource.push(new ProxyEntity(proxy.toEntity()));
```