Reconcile custom Catalog entities with API Objects, add License (#744)
This update adds the missing LicenseEntity for adding licenses to the Catalog.

It also establishes all the typesets for expected shapes of catalog entities
of all the custom types.

It removes all the creation-related context menu items since those aren't
part of MVP (except maybe "create cluster", but let's see if we get there).

And finally, it adds a `toEntity()` method to each of the API types
(Cluster, Credential, Proxy, SshKey, License, Namespace) to use when
adding items to the Catalog:

```
catalogSource.push(new ProxyEntity(proxy.toEntity()));
```

Finally, finally, the `ClusterStore` has been entirely removed. There is no
possible migration strategy for it from v3 to v4 because in v4, we're stopping
at the namespace and pulling all clusters from a synced NS into Lens. We
couldn't, for example, somehow sync only select clusters that came from
the `ClusterStore`.