Rename ExtendedCloud to DataCloud
It's been bugging me for a while. I think DataCloud is at least
marginally more meaningful than ExtendedCloud. A DataCloud is an
object that uses a Cloud's connection to get its data.

Also, "Extended" was sounding too much like ExtendedCloud _extends from_
Cloud, which isn't the case. Maybe DataCloud isn't that much better.
But it's shorter to write. That should be good.