Rename Kube-related classes/var names to be more aligned with Kube API
The [Kube API](https://kubernetes.io/docs/reference/using-api/api-concepts/#standard-api-terminology)
uses specific terminology which the code was not aligned with. Since
MCC's API is really the Kube API with MCC extensions, the code that
deals with the API should be more closely aligned with Kube API
concepts so that it's more familiar to someone who's familiar with
Kubernetes.