[PRODX-22269] SyncManager efficient updates
The SM is now much smarter than it was before because it leverages
the Kubernetes-provided `resourceVersion` as a way to detect if
a resource has changed since the last time it was synced.

The result is that only Catalog entities that are older than their
related resource get updated. Everything else that hasn't changed
just gets skipped. That greatly reduces churn.

Also:

- Deletes are now supported, cleaning-up unused kubeconfig files
  (deletes meaning a namespace is no longer synced, or a Cloud
  is removed).
- Fixed bug where we'd get an empty screen with a spinner at the
  top/left right after adding a new (first) Cloud.
- Fixed bug where a removed Cloud would remain in the SyncView after
  being removed.