[PRODX-22307] Relate resources to clusters, add machines, add apiStatus
Relates other resources (SSH keys, credentials, proxies, licenses, machines)
to the clusters that use them.

Fetches all machines in all synced namespaces so that we can determine
a cluster's master and worker node counts, as well as it license
(since RHEL licenses are related to OpenStack cluster machines, not
the clusters themselves).

Adds a new `apiStatus` property to the Cluster Entity Model which
will provide a value for the "MCC Status" field in "View Details".
This is determined via "status conditions" on the cluster.

Also updates labels like this:
- Renames the much too long "managementCluster" label to "mgmt-cluster".
- Renames "sshKey" label to "ssh-key" as kebabcase seems more standard
  for label/tag names.
- Adds new "is-mgmt-cluster=true" label to all mgmt clusters.
- All clusters, whether mgmt or not, now get the "mgmt-cluster=name",
  "project=name", and any applicable of "ssh-key", "credential",
  "proxy", "license" labels since they all apply to all types of
  clusters.

Also fixed a small bug with incorrect destructuring of `unknownValue()`
string function in catalogEntityDetails.tsx leading to a crash if the
function is called for an unknown value.