[PRODX-22753] Fix bug when 2 clouds, disconnected, and connect just 1 (#801)
The result before the fix was that all entities related to the still-
disconnected Cloud would get removed, when they should remain until
the Cloud is reconnected and synced, or removed outright.

The result after the fix is that only the entities and models related
to the Cloud being updated are touched in the Catalog and SyncStore.
All others are left alone. So the items for the still-disconnected
Cloud remain untouched.