[PRODX-22830] BIG REFACTOR: No longer fetch on Renderer thread
This large refactoring ultimately addresses 2 main issues:

1.  Performance issues during sync caused by overwhelming ENZI
    which runs behind the scenes on every mgmt cluster with
    ~12 requests per namespace (synced or not, since we were
    doing discovery of clusters, credentials, and keys even on
    non-synced ones), which could quickly add up to hundreds
    for mgmt clusters with lots of namespaces. And then we were
    __doubling__ these requests by doing them simultaneously on
    the MAIN and RENDERER threads. Ouch.
2.  Endless ping-pong via cloud-store.json (CloudStore class)
    between MAIN and RENDERER thread updates to Clouds as various
    state would change.

What we now have is UI code (i.e. RENDERER thread) that uses Clouds,
not DataClouds (except when adding a new mgmt cluster and needing
to pre-fetch its namespaces to enable the initial sync selection
from the user).

The Cloud class now has a `namespaces` array property which contains
CloudNamespace class instances, one for each known namespace
(synced or not), along with simple counts of how many resources
of what type it contains. These are updated on the MAIN thread
only via use of DataClouds by the SyncManager. RENDERER gets the
updated data via CloudStore updates when MAIN writes to
cloud-store.json on disk.

The Cloud instances now have new `loading` and `fetching` properties,
and these, along with `connecting` and `connectError` (none of which
are stored in cloud-store.json) are updated from MAIN to RENDERER
only via IPC broadcasts to avoid trying to use the CloudStore (which
would lead to more ping-ponging).

The RENDERER's DataCloudProvider is now just a CloudProvider.

The logic in CloudProvider and SyncManager for dealing with CloudStore
updates has been drastically simplified. The CloudStore is also much
more precise about which Cloud change events it cares about, which
is now specific to what thread its on since only certain types of
changes are expected on a specific thread.

SSO requests to connect to Clouds are now initiated by the SyncManager
on MAIN instead of somewhere on RENDERER.

The Namespace API resource class now has a _preview_ mode where its
`*Count` properties return values from numbers instead of using the
length of a related array property, for when we're adding a new
mgmt cluster and we discover namespaces without fetching all the
rest of the data (clusters, credentials, etc.).