[PRODX-22469] Implement Kube watch API support after full data fetch (#829)
The idea was to leverage Kube's watch API[^1] to make fetches more
efficient instead of just a dumb 5 minute poll, especially, perhaps,
when syncing with multiple Clouds.

Implementation is done, but STUCK on a hard-to-debug or figure-out
issue, as documented with `// TODO[PRODX-22469]` comments in
DataCloud.js and SyncManager.js

For now, DataCloud instance continue to use polling. Besides, apparently
Kube watches expire after 5 minutes and a full data fetch has to be
done again to get a new `resourceVersion` for the collection anyway,
so it may not be worth spending more time on this and just ripping it
out. Search for all the TODO comments to know what to rip out.

Aside from this:

- Fixed a bug with wrong event name "LOADING_CHANGE" -> "LOADED_CHANGE"
- Fixed a bug where `extractBodyMethod` wasn't being passed all the way
  down into `netUtil.request()`
- Fixed unhandled failed promise when destroying a Cloud (related to
  attempting to log out on the RENDERER thread without having a config)
- Added new `EventDispatcher.sendEvent()` method which guarantees that
  listeners will see every dispatched event instead of just the most
  recent one with most recent params via `dispatchEvent()`

[^1]: https://kubernetes.io/docs/reference/using-api/api-concepts/#efficient-detection-of-changes