Add friendly/sticky error message for legacy protocol requests
Instead of just no longer adding the handlers, put the handlers
back and show a nice, sticky error notification. What Lens shows
by default is really awful, and doesn't stick either. So users
are left confused when things don't work as expected (typically,
they have upgrade their extension version to 5.0 but are still
using it with MCC < 2.19 and expecting it to work like it used to).