[CCLEX-183] Handle not having lma; show enabled state in details
The `lma` data of a cluster entity will be `null` if the cluster
isn't configured with StackLight. The code needs to handle this
case.

Also, now that we're not showing/hiding the LMA panel, it seems
good to include an 'Enabled: Yes/No' field to indicate whether
StackLight is configured on the cluster or not, since it might
not be obvious just with the '--' values for URLs. For example,
are all the URLs not available because SackLight is disabled, or
because those services aren't available for some reason...?