[CCLEX-178] Add metrics API module with mock API responses
Since we're blocked on PRODX-28459 in terms of accessing the Prometheus
API on a given cluster (cloud), this commit includes a
`./__tests__/metricPromMock.js` module that currently takes care
of fulfilling Prometheus API requests made by the new metricApi.js
module.

The metricApi.js module seeks to abstract the details of calling
the Prometheus API by exposing high level functions like
`getCpuMetrics()`, `getMemoryMetrics()`, and `getDiskMetrics()`
which, internally, execute multiple Prometheus queries in order
to gather the necessary data.

We should be able to build the Health panel based on this module
and the fake data for now until we can get rid of the mock layer
and reach Prometheus directly.