[CCLEX-74] REMOVE kube watches implementation
It was added last April, but couldn't be enabled and tested because
of an issue with Node and extracting the body of the watch payload:

  Disabling watches until a solution can be found to the behavior where a watch
  request's response comes back almost immediately as status 200, but then
  attempting to extract the body with `response.text()` in
  `netUtils.js#tryExtractBody()` NEVER resolves. It just hangs.

  I thought this might be similar to this issue: https://github.com/node-fetch/node-fetch/issues/665
  And solution (for node-fetch 2.x): https://github.com/node-fetch/node-fetch#custom-highwatermark

  But, alas, that didn't help at all. No clue what's going on. Works great in
  Postman, so clearly, the issue is with how we're handling these types of
  long-poll requests in `netUtil.js#request()`, but not sure what to do.

Based on kube docs, watches expire after 5 minutes (server just doesn't preserve
historical record of changes for longer than 5 minutes by default because it's
very data-intensive, I presume) and once they expire, the API documentation says
a full collection fetch must be performed (i.e. get all data from it) in order
to get a new `resourceVersion` and keep watching from there; repeat...

So if watches expire every 5 minutes anyway, we're still doing full pulls every
5 minutes, so the only thing the great complexity of watches buys us is
potential finer-grained, more immediate updates in between full fetches, which
seems unnecessary.

Watches were never enabled, and we've been doing fine without them. Furthermore,
we've discovered that the person using Lens is normally not also the same person
using MCC in the browser, so the likelihood that someone will have both open at
the same time, expecting near instantaneous updates in Lens in response to
changes in MCC, will be very rare. And in that case, they can just use the
"Sync now" feature.

Finally, we’re already successfully using the `resourceVersion` associated with
individual resources in order to skip some updates and make code more efficient.

So based on all of the above, I don't think it's worth carrying the extra weight
of all this untested/untestable/unusable code around in case it _might_ work
some day, and also trying to maintain it without being able to test the changes.