[CCLEX-186] Add cluster updates to API and cluster entities
This adds an `updates` property on Cluster objects and cluster
entities containing cluster/machine deployment/upgrade history,
if available.

It also renames the 'Performance' tab to 'Updates' since we won't
have a Performance tab anymore, but we're adding the Updates tab.

Since we need MCC v2.22 to get the new `stages[].status` field in
`ClusterDeploymentStatus`, `ClusterUpgradeStatus`, and
`MachineUpgradeStatus` (machines don't get deployments, only
upgrades), fetching updates is behind a new Webpack-enabled
`FEAT_CLUSTER_PAGE_UPDATES_ENABLED` switch which can be set
on the command line to opt in to seeing the Updates tab when
running in Dev mode (or building Prod with the flag).