Align with 'release' on upgrade objects in future API work
PRODX-28727 will rename the 'toRelease' property on upgrade objects
to 'release' in order to align it with existing 'release' property
on deployment objects to make things more consistent.

This change aligns with that in advance of it happening in so far
as our own code is concerned.

Also renames the special flag to use 'HISTORY' instead of 'UPDATES'
so it's more clear what it's related to.