[CCLEX-179] Add support for offline access tokens
Now cluster access tokens can optionally be generated as offline
(i.e. long-lived) tokens on a per-mgmt cluster basis.

This works around (well, _delays_ for a long time)
https://github.com/lensapp/lens/issues/6966

NOTE: Also has a partial implementation of the new 'trust host'
feature that was started at the same time as 'offline access', but
which is not currently exposed to the user.