[CCLEX-194] Surface better connection error messages
When adding a new mgmt cluster, if a connection error occurs, we now
parse it to see if it's a certificate error (self-signed certs), a
connection error, or some other unknown error, and display the friendly
error message in an error notification popup that sticks to the screen
until the user manually dismisses it (instead of having it stay visible
only for 1-2 seconds and always be a generic message).

When reconnecting a disconnected mgmt cluster, if there's a connection
error, we now show a little error icon next to the "Disconnected"
status, with a friendly, more helpful connection error message in
a tooltip on hover over the error icon.

Also:
- Added better disabled state support to TriStateCheckbox
- Made sure that Connect button (when adding a new mgmt cluster) remains
  disabled if any of the required field validators fail