[CCLEX-179] Add trust host cloud option (#1291)
Users now have the option to 'trust this host' when adding a new mgmt
cluster (only when adding, not in Selective Sync view) in order to
get around issues with self-signed certificates for which the
Lens command line option `LEX_CC_UNSAFE_ALLOW_SELFSIGNED_CERTS`
was previously used.

With this change, the command line option has been removed.

The configuration setting is persisted in the cloud-store.json and
restored on Lens restart.

To untrust a host, remove the cloud and adding it again without the
option turned on. This is because we assume that if a host must be
trusted, it likely always needs to be trusted. The inverse is not
true: If a host was trusted and is no longer verifiable, something
may be aloof and the problem should be rectified instead of making
it easy to just trust and untrust hosts.

Also:
- Fixed TriStateCheckbox help text not using disabled color.