[CCLEX-179] Add 'Configure trusted hosts' migration dialog
Since the `LEX_CC_UNSAFE_ALLOW_SELFSIGNED_CERTS` flag has been removed,
but users will still be using it after installing v5.5.0, if the flag
is set, and the user is syncing at least one untrusted cloud with a
secure (https) connection, a special dialog will be displayed showing
a checkbox per secure, untrusted cloud, allowing the user an opportunity
to trust the clouds that need to be trusted.

This dialog is only shown once per Lens session.

Also added new Modal component, since the Lens Dialog component is
not designed to be used on its own, and Lens doesn't provide a
generic dialog component like that which was needed.

Also updated the 'Add mgmt cluster' view to disable the 'trust this
host' option when the URL is not https, and restricted the mgmt
cluster URL to http or https only since those are the protocols we
expect for an MCC instance.