[CCLEX-188] Add MCC release info to all Catalog entities
First, the active KAAS_RELEASE object had to be fetched, which led to
some refactoring of the API Client classes.

Then it had to be used to compare the active version with a sought
version, hence the new `Release` API resource type class.

Then it had to get all the down into Catalog entities. But since the
release is fetched on every sync (since it could change from one
sync to another), the release is DataCloud-related data, not
Cloud-related data, and all API resource type classes were requiring
Cloud instances, not DataCloud instances. So they all had to be
refactored to accept a DataCloud instead.

At this point, everything works again like it used to. There are
no functional changes other than:

1. The release is always fetched.
2. Instead of looking at the `FEAT_CLUSTER_PAGE_HISTORY_ENABLED`
   Webpack flag during DataCloud fetch to determine if we should
   fetch resource updates, now we use the Release instance and
   fetch resource updates if the active release is `>=2.22`. The
   rest of the history-related code still uses the Webpack flag.
3. The release info is included in all Catalog entities.

Finally, the `semver` dependency was added to do the semver
comparisons with the release version.