ExtensionStore is no longer a Singleton in Lens since some 6.4.x release
- https://github.com/lensapp/lens/blob/master/packages/core/src/extensions/extension-store.ts
- https://github.com/lensapp/lens/pull/6690 ("Make base store non Singleton")
- https://github.com/lensapp/lens/issues/4573 ("Elimination of global shared state")

The `createInstance()` and `getInstance()` methods had been deprecated in the Lens
Extensions API, so they have been eliminated from our code base also. But we're still
using "global shared state" Cloud and Sync store instances. It's just that Lens isn't
the one creating and holding the instance anymore, we are.